@@ -1,7 +1,5 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
-import 'package:multiplatform_widgets/multiplatform_widgets.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class AboutScreen extends StatelessWidget {
@@ -10,7 +8,11 @@ class AboutScreen extends StatelessWidget {
   final String history;
   final String version;
 
-  AboutScreen({this.about, this.history, this.version});
+  AboutScreen(
+      {super.key,
+      required this.about,
+      required this.history,
+      required this.version});
 
   final controller = ScrollController();
 
@@ -18,47 +20,25 @@ class AboutScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     // String about;
 
-    return MpScaffold(
-      appBar: MpAppBar(
-          title: Text('About'),
-          button: MpLinkButton(
-            label: 'Licenses',
-            onPressed: () => showLicensePage(
-                context: context,
-                applicationVersion: this.version,
-                applicationLegalese: '© Denis Filonov'),
-          )),
+    return Scaffold(
+      appBar: AppBar(title: const Text('About'), actions: [
+        MaterialButton(
+          child: const Text('Licenses'),
+          onPressed: () => showLicensePage(
+              context: context,
+              applicationVersion: version,
+              applicationLegalese: '© Denis Filonov'),
+        )
+      ]),
       body: SafeArea(
         child: Markdown(
-          data: about + '\n' + history,
+          data: '$about\n$history',
           controller: controller,
           selectable: false,
           styleSheetTheme: MarkdownStyleSheetBaseTheme.platform,
-          onTapLink: (href) => launch(href),
+          onTapLink: (text, href, title) => launchUrl(Uri.parse(href!)),
         ),
       ),
     );
   }
 }
-
-// Container(
-//   height: 260.0,
-//   // constraints: BoxConstraints(maxWidth: 800.0),
-//   child: FutureBuilder(
-//       future: rootBundle.loadString("text/$myLocale/about.md"),
-//       builder:
-//           (BuildContext context, AsyncSnapshot<String> snapshot) {
-//         if (snapshot.hasData) {
-//           return Markdown(
-//             data: snapshot.data,
-//             controller: controller,
-//             selectable: false,
-//             styleSheetTheme: MarkdownStyleSheetBaseTheme.platform,
-//             onTapLink: (href) => launch(href),
-//           );
-//         }
-//         return Center(
-//           child: CircularProgressIndicator(),
-//         );
-//       }),
-// ),
