@@ -1,7 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
-import 'screens/about_screen.dart';
 import 'screens/timer/set_screen.dart';
 import 'screens/timer/timer_screen.dart';
 import 'screens/timer/train_screen.dart';
@@ -11,7 +10,6 @@ import 'screens/timer/train_screen.dart';
 /// Material light theme
 final ThemeData kMaterialLight = ThemeData(
   primarySwatch: Colors.blue,
-  buttonColor: Colors.blue,
   brightness: Brightness.light,
   visualDensity: VisualDensity.adaptivePlatformDensity,
 );
@@ -35,8 +33,7 @@ final CupertinoThemeData kCupertinoTheme = CupertinoThemeData(
 final kInitialRoute = TrainScreen.id;
 
 final Map<String, Widget Function(BuildContext)> kAppRoutes = {
-  TrainScreen.id: (context) => TrainScreen(),
-  SetScreen.id: (context) => SetScreen(),
-  TimerScreen.id: (context) => TimerScreen(),
-  AboutScreen.id: (context) => AboutScreen(),
+  TrainScreen.id: (context) => const TrainScreen(),
+  SetScreen.id: (context) => const SetScreen(),
+  TimerScreen.id: (context) => const TimerScreen(),
 };
