@@ -7,7 +7,8 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
+// ignore_for_file:unnecessary_string_interpolations, unnecessary_string_escapes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -20,14 +21,15 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'en';
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "about" : MessageLookupByLibrary.simpleMessage("About"),
-    "from" : MessageLookupByLibrary.simpleMessage(" from "),
-    "rest" : MessageLookupByLibrary.simpleMessage("Rest time (seconds)"),
-    "restButton" : MessageLookupByLibrary.simpleMessage("Rest"),
-    "sets" : MessageLookupByLibrary.simpleMessage("Sets"),
-    "setsText" : MessageLookupByLibrary.simpleMessage("Perfom the set, and when you\'re done, press the button."),
-    "start" : MessageLookupByLibrary.simpleMessage("Start"),
-    "title" : MessageLookupByLibrary.simpleMessage("Sets Timer")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "about": MessageLookupByLibrary.simpleMessage("About"),
+        "from": MessageLookupByLibrary.simpleMessage(" from "),
+        "rest": MessageLookupByLibrary.simpleMessage("Rest time (seconds)"),
+        "restButton": MessageLookupByLibrary.simpleMessage("Rest"),
+        "sets": MessageLookupByLibrary.simpleMessage("Sets"),
+        "setsText": MessageLookupByLibrary.simpleMessage(
+            "Perfom the set, and when you\'re done, press the button."),
+        "start": MessageLookupByLibrary.simpleMessage("Start"),
+        "title": MessageLookupByLibrary.simpleMessage("Sets Timer")
+      };
 }
