@@ -7,7 +7,8 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
+// ignore_for_file:unnecessary_string_interpolations, unnecessary_string_escapes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -20,14 +21,15 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'ru_RU';
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "about" : MessageLookupByLibrary.simpleMessage("О приложении"),
-    "from" : MessageLookupByLibrary.simpleMessage(" из "),
-    "rest" : MessageLookupByLibrary.simpleMessage("Время отдыха в секундах"),
-    "restButton" : MessageLookupByLibrary.simpleMessage("Отдых"),
-    "sets" : MessageLookupByLibrary.simpleMessage("Количество подходов"),
-    "setsText" : MessageLookupByLibrary.simpleMessage("Выполните подход, и когда закончите, нажмите кнопку."),
-    "start" : MessageLookupByLibrary.simpleMessage("Старт"),
-    "title" : MessageLookupByLibrary.simpleMessage("Таймер подходов")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "about": MessageLookupByLibrary.simpleMessage("О приложении"),
+        "from": MessageLookupByLibrary.simpleMessage(" из "),
+        "rest": MessageLookupByLibrary.simpleMessage("Время отдыха в секундах"),
+        "restButton": MessageLookupByLibrary.simpleMessage("Отдых"),
+        "sets": MessageLookupByLibrary.simpleMessage("Количество подходов"),
+        "setsText": MessageLookupByLibrary.simpleMessage(
+            "Выполните подход, и когда закончите, нажмите кнопку."),
+        "start": MessageLookupByLibrary.simpleMessage("Старт"),
+        "title": MessageLookupByLibrary.simpleMessage("Таймер подходов")
+      };
 }
