@@ -1,13 +1,9 @@
-import 'dart:io' show Platform;
-import 'package:flutter/foundation.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-// Localization
 import 'package:flutter_localizations/flutter_localizations.dart';
-import 'config.dart';
-import 'package:setstimer/generated/l10n.dart';
 import 'package:provider/provider.dart';
 
+import 'config.dart';
+import 'generated/l10n.dart';
 import 'models/set_rest.dart';
 
 void main() {
@@ -16,18 +12,18 @@ void main() {
       providers: [
         ChangeNotifierProvider(create: (context) => SetRestData()),
       ],
-      child: (!kIsWeb && (Platform.isMacOS || Platform.isIOS))
-          ? AppCupertino()
-          : AppMaterial(),
+      child: const App(),
     ),
   );
 }
 
-class AppMaterial extends StatelessWidget {
+class App extends StatelessWidget {
+  const App({super.key});
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      localizationsDelegates: [
+      localizationsDelegates: const [
         S.delegate,
         GlobalMaterialLocalizations.delegate,
         GlobalWidgetsLocalizations.delegate,
@@ -37,25 +33,6 @@ class AppMaterial extends StatelessWidget {
       onGenerateTitle: (BuildContext context) => S.of(context).title,
       theme: kMaterialLight,
       darkTheme: kMaterialDark,
-      initialRoute: kInitialRoute,
-      routes: kAppRoutes,
-    );
-  }
-}
-
-class AppCupertino extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return CupertinoApp(
-      localizationsDelegates: [
-        S.delegate,
-        GlobalMaterialLocalizations.delegate,
-        GlobalWidgetsLocalizations.delegate,
-        GlobalCupertinoLocalizations.delegate,
-      ],
-      supportedLocales: S.delegate.supportedLocales,
-      onGenerateTitle: (BuildContext context) => S.of(context).title,
-      theme: kCupertinoTheme,
       initialRoute: kInitialRoute,
       routes: kAppRoutes,
     );
